Mask Code setting. Panel Arming even migration.
Added option to show or mask alarm codes in logs. Fixed bug in list of zones bypassed. Panel Arming Events fields are disabled. The default for User was set to select User to for migration of User Actions. Added code to make Panel Arming migration to User Actions easier for users - preset to Any User.