Checking Zones and Battery Restore State
Added a new Alarm Zone device state Last Battery Restore. Zone numbers that are not valid (not in the range of 1-99) within Keypad Messages and will now log a warning versus an error in the Indigo Event Log. Zone numbers in Alarm Panel messages that are not defined as an Alarm Zone Device in Indigo will log a new warning messages in the Indigo Event Log.
Added two new User Events: Walk Test Mode On and Walk Test Mode Off.